backend/requirements.txt
APScheduler==3.9.1
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
boto3==1.22.8
botocore==1.25.8
CacheControl==0.12.11
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
Django==4.0.4
django-apscheduler==0.6.2
django-cors-headers==3.11.0
django-storages==1.12.3
djangorestframework==3.13.1
drf-yasg==1.20.0
firebase-admin==5.2.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.46.0
grpcio-status==1.46.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httplib2==0.20.4
httpx==0.13.3
hyperframe==5.2.0
idna
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
jmespath==1.0.0
korcen==0.0.8
MarkupSafe==2.1.1
msgpack==1.0.3
mysqlclient==2.1.0
numpy==1.22.3
outcome==1.1.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
proto-plus==1.20.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
trio==0.20.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
wsproto==1.1.0
